minData Update qPCRadmin.R
Updated minData function to be able to handle .csv, .xls, and .xlsx files as input. In addition, it can handle any organizational model of the data as long as the first column for the results begins in 'Well'. 
Previous version needed that header column to be at a precise location for SYBR or TAQMAN files.